Test non-blocking behavior with updated go-cs-bouncer; removed mock lapi
