+++ main.go
@@ -19,16 +19,14 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/sirupsen/logrus/hooks/writer"
 
+	"github.com/crowdsecurity/crowdsec/pkg/models"
 	"github.com/crowdsecurity/cs-custom-bouncer/pkg/version"
 	csbouncer "github.com/crowdsecurity/go-cs-bouncer"
-	"gopkg.in/tomb.v2"
 )
 
 const (
 	name = "crowdsec-custom-bouncer"
 )
-
-var t tomb.Tomb
 
 func termHandler(sig os.Signal, custom *customBouncer) error {
 	if err := custom.ShutDown(); err != nil {
@@ -61,6 +59,38 @@ func HandleSignals(custom *customBouncer) {
 	log.Infof("Shutting down custom-bouncer service")
 	os.Exit(code)
 }
+
+func deleteDecisions(custom *customBouncer, decisions []*models.Decision) {
+	if len(decisions) == 1 {
+		log.Infof("deleting 1 decision")
+	} else {
+		log.Infof("deleting %d decisions", len(decisions))
+	}
+	for _, d := range decisions {
+		if err := custom.Delete(d); err != nil {
+			log.Errorf("unable to delete decision for '%s': %s", *d.Value, err)
+			continue
+		}
+		log.Debugf("deleted '%s'", *d.Value)
+	}
+}
+
+
+func addDecisions(custom *customBouncer, decisions []*models.Decision) {
+	if len(decisions) == 1 {
+		log.Infof("adding 1 decision")
+	} else {
+		log.Infof("adding %d decisions", len(decisions))
+	}
+	for _, d := range decisions {
+		if err := custom.Add(d); err != nil {
+			log.Errorf("unable to insert decision for '%s': %s", *d.Value, err)
+			continue
+		}
+		log.Debugf("Adding '%s' for '%s'", *d.Value, *d.Duration)
+	}
+}
+
 
 func main() {
 	var err error
@@ -104,11 +134,11 @@ func main() {
 
 	custom, err := newCustomBouncer(config)
 	if err != nil {
-		log.Fatalf(err.Error())
+		log.Fatal(err)
 	}
 
 	if err := custom.Init(); err != nil {
-		log.Fatalf(err.Error())
+		log.Fatal(err)
 	}
 
 	bouncer := &csbouncer.StreamBouncer{}
@@ -116,19 +146,22 @@ func main() {
 
 	err = bouncer.ConfigReader(bytes.NewReader(configBytes))
 	if err != nil {
-		log.Errorf("unable to configure bouncer: %s", err)
-		return
+		log.Fatalf("unable to configure bouncer: %s", err)
 	}
 
 	if err := bouncer.Init(); err != nil {
-		log.Error(err.Error())
-		return
+		log.Fatal(err)
 	}
 	cacheResetTicker := time.NewTicker(config.CacheRetentionDuration)
+
+	ctx, cancel := context.WithCancel(context.Background())
+
 	go func() {
-		bouncer.Run()
-		t.Kill(fmt.Errorf("stream init failed"))
+		bouncer.Run(ctx)
+		log.Errorf("stream init failed")
+		cancel()
 	}()
+
 	if config.PrometheusConfig.Enabled {
 		listenOn := net.JoinHostPort(
 			config.PrometheusConfig.ListenAddress,
@@ -147,50 +180,49 @@ func main() {
 		go func() {
 			log.Infof("Serving metrics at %s", listenOn+"/metrics")
 			log.Error(promServer.ListenAndServe())
+			// don't need to cancel context here, prometheus is not critical
 		}()
 	}
 	if config.FeedViaStdin {
-		t.Go(
-			func() error {
-				f := func() error {
-					log.Debugf("Starting binary %s %s", config.BinPath, config.BinArgs)
-					c := exec.Command(config.BinPath, config.BinArgs...)
-					s, err := c.StdinPipe()
-					if err != nil {
-						return err
-					}
-					custom.binaryStdin = s
-					if err := c.Start(); err != nil {
-						return err
-					}
+		go func() error {
+			defer cancel()
 
-					return c.Wait()
+			f := func() error {
+				log.Debugf("Starting binary %s %s", config.BinPath, config.BinArgs)
+				c := exec.CommandContext(ctx, config.BinPath, config.BinArgs...)
+				s, err := c.StdinPipe()
+				if err != nil {
+					return err
 				}
-				var err error
-				if config.TotalRetries == -1 {
-					for {
-						err := f()
-						log.Errorf("Binary exited: %s", err)
-					}
-				} else {
-					for i := 0; i <= config.TotalRetries; i++ {
-						err = f()
-						log.Errorf("Binary exited (retry %d/%d): %s", i, config.TotalRetries, err)
-					}
+				custom.binaryStdin = s
+				if err := c.Start(); err != nil {
+					return err
 				}
-				log.Error("maximum retries exceeded for binary. Exiting")
-				t.Kill(err)
-				return err
-			},
-		)
 
+				return c.Wait()
+			}
+			var err error
+			if config.TotalRetries == -1 {
+				for {
+					err := f()
+					log.Errorf("Binary exited: %s", err)
+				}
+			} else {
+				for i := 1; i <= config.TotalRetries; i++ {
+					err = f()
+					log.Errorf("Binary exited (retry %d/%d): %s", i, config.TotalRetries, err)
+				}
+			}
+			log.Error("maximum retries exceeded for binary. Exiting")
+			return err
+		}()
 	}
 
-	t.Go(func() error {
-		log.Printf("Processing new and deleted decisions . . .")
+	go func() {
+		log.Infof("Processing new and deleted decisions . . .")
 		for {
 			select {
-			case <-t.Dying():
+			case <-ctx.Done():
 				log.Infoln("terminating bouncer process")
 				if config.PrometheusConfig.Enabled {
 					log.Infoln("terminating prometheus server")
@@ -198,30 +230,15 @@ func main() {
 						log.Errorf("unable to shutdown prometheus server: %s", err)
 					}
 				}
-				return nil
+				return
 			case decisions := <-bouncer.Stream:
-				log.Infof("deleting '%d' decisions", len(decisions.Deleted))
-				for _, decision := range decisions.Deleted {
-					if err := custom.Delete(decision); err != nil {
-						log.Errorf("unable to delete decision for '%s': %s", *decision.Value, err)
-					} else {
-						log.Debugf("deleted '%s'", *decision.Value)
-					}
-
-				}
-				log.Infof("adding '%d' decisions", len(decisions.New))
-				for _, decision := range decisions.New {
-					if err := custom.Add(decision); err != nil {
-						log.Errorf("unable to insert decision for '%s': %s", *decision.Value, err)
-					} else {
-						log.Debugf("Adding '%s' for '%s'", *decision.Value, *decision.Duration)
-					}
-				}
+				deleteDecisions(custom, decisions.Deleted)
+				addDecisions(custom, decisions.New)
 			case <-cacheResetTicker.C:
 				custom.ResetCache()
 			}
 		}
-	})
+	}()
 
 	if config.Daemon {
 		sent, err := daemon.SdNotify(false, "READY=1")
@@ -231,7 +248,6 @@ func main() {
 		go HandleSignals(custom)
 	}
 
-	if err := t.Wait(); err != nil {
-		log.Errorf("process return with error: %s", err)
-	}
+	// Wait for the context to be cancelled
+	<-ctx.Done()
 }
